% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrumar_solucao.R
\name{arrumar_solucao}
\alias{arrumar_solucao}
\title{Arrumar solução do otimizador}
\usage{
arrumar_solucao(res)
}
\arguments{
\item{res}{Objeto retornado pela função \code{\link{otimizar}}.}
}
\value{
lista
}
\description{
Arrumar solução do otimizador
}
