% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_inputs.R
\name{calcular_inputs}
\alias{calcular_inputs}
\title{Calcular inputs do modelo}
\usage{
calcular_inputs(
  con,
  parametros,
  regiao_escopo,
  populacao_minima = 10000,
  rotas = c("dens", "biocomb", "bioeletr", "bioeletr_pec", "biocomb_pec", "cofiring"),
  produtos_selecionados = NULL,
  backend = c("disco", "ram"),
  sazonalidade = TRUE
)
}
\arguments{
\item{con}{Conexão com o banco de dados do otimizador.}

\item{parametros}{Tabela de parâmetros.}

\item{regiao_escopo}{Vetor com siglas de estados brasileiros ou a string
"Brasil" para considerar todo o país como regiao escopo.}

\item{populacao_minima}{População mínima que um município deve possuir
para que ele possa ser uma sede.}

\item{rotas}{Vetor de rotas.}

\item{produtos_selecionados}{Vetor com o id dos produtos que vão entrar no
cálculo do energético.}

\item{backend}{As manipulações da base devem ser feitas em disco, usando
SQL, ou em RAM, usando dplyr? Opções: 'disco' (padrão) ou 'ram'.}

\item{sazonalidade}{A sazonalidade deve ser considerada na otimização?
Opções: TRUE (sim, argumento padrão) ou FALSE (não)}
}
\value{
Uma lista com os inputs de cada rota.
}
\description{
Calcular inputs do modelo
}
