% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparar_tabela_resultados.R
\name{calcular_porc_demanda_satisfeita}
\alias{calcular_porc_demanda_satisfeita}
\title{calcula a porcentagem da demanda satisfeita}
\usage{
calcular_porc_demanda_satisfeita(resultado, demanda_total)
}
\arguments{
\item{resultado}{saída da função 'arrumar_solucao'}

\item{demanda_total}{tabela com a demanda total de cada rota}
}
\description{
calcula a porcentagem da demanda satisfeita
}
