% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_potencia.R
\name{calcular_potencia}
\alias{calcular_potencia}
\title{Calcular a potência em MW}
\usage{
calcular_potencia(energia, fator_disp)
}
\arguments{
\item{energia}{Energia em MWh.}

\item{fator_disp}{Fator de disponibilidade.}
}
\value{
vetor de tamanho 1
}
\description{
Calcular a potência em MW
}
