% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dados_geo}
\alias{dados_geo}
\title{Polígonos de Estados e Municípios}
\format{
An object of class \code{list} of length 27.
}
\usage{
dados_geo
}
\description{
Dados sobre Estados e Municípios que possibilitam a construção de mapas.
}
\keyword{datasets}
