% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizar_solucao.R
\name{ec_regioes}
\alias{ec_regioes}
\title{echart com mapa das regiões}
\usage{
ec_regioes(tab, regiao_escopo, sedes = NULL, title = "")
}
\arguments{
\item{tab}{tabela de solução do otimizador.}

\item{regiao_escopo}{Região escopo da otimização.}

\item{sedes}{Sedes específicas a serem mostradas no mapa.}

\item{title}{Título do mapa.}
}
\value{
ggplot com o mapa.
}
\description{
echart com mapa das regiões
}
