% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estados_br}
\alias{estados_br}
\title{Estados brasileiros}
\format{
An object of class \code{character} of length 27.
}
\usage{
estados_br
}
\description{
Vetor de texto com a sigla dos estados brasileiros
}
\keyword{datasets}
