% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizar_solucao.R
\name{gg_regioes}
\alias{gg_regioes}
\title{ggplot com mapa das regiões}
\usage{
gg_regioes(
  tab,
  regiao_escopo,
  sedes = NULL,
  title = "",
  ref = "Sem aproveitamento de resíduo"
)
}
\arguments{
\item{tab}{tabela de solução do otimizador.}

\item{regiao_escopo}{Região escopo da otimização.}

\item{sedes}{Sedes específicas a serem mostradas no mapa.}

\item{title}{Título do mapa.}

\item{ref}{Referência para municípios sem solução.}
}
\value{
ggplot com o mapa.
}
\description{
ggplot com mapa das regiões
}
